chore: move swr examples to nextjs example
