+++ packages/nextjs-testapp/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@wundergraph/nextjs": "workspace:*",
     "@wundergraph/sdk": "workspace:*",
+    "@wundergraph/swr": "workspace:*",
     "next": "^12.1.6",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
