+++ packages/sdk/src/internal/index.ts
@@ -1,7 +1,7 @@
 /**
  * This API is exported but only intended for internal use. Don't rely on it.
  */
-export { visitJSONSchema } from '../codegen/index';
+export { visitJSONSchema } from '../codegen/jsonschema';
 
 export { serialize } from '../utils';
