--- packages/vite-testapp/.graphqlconfig
@@ -1,19 +1,0 @@
-{
-  "projects": {
-    "api": {
-      "name": "api",
-      "schemaPath": ".wundergraph/generated/wundergraph.api.schema.graphql",
-      "extensions": {
-        "endpoints": {
-          "api": {
-            "introspect": false,
-            "url": "http://localhost:9991/api/main/graphql",
-            "headers": {
-              "user-agent": "WunderGraph Client"
-            }
-          }
-        }
-      }
-    }
-  }
-}
