--- packages/vite-testapp/.wundergraph/jsonplaceholder.v1.yaml
@@ -1,155 +1,0 @@
-openapi: 3.0.0
-info:
-  title: jsonplaceholder
-  version: '1.0'
-servers:
-  - url: 'https://jsonplaceholder.typicode.com'
-paths:
-  /posts:
-    get:
-      summary: Your GET endpoint
-      tags: []
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/Post'
-      operationId: posts
-  '/posts/{postID}/comments':
-    parameters:
-      - schema:
-          type: string
-        name: postID
-        in: path
-        required: true
-    get:
-      summary: Your GET endpoint
-      tags: []
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/Comment'
-      operationId: postComments
-  /users:
-    get:
-      summary: Your GET endpoint
-      tags: []
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/User'
-      operationId: users
-  '/users/{userID}/posts':
-    parameters:
-      - schema:
-          type: string
-        name: userID
-        in: path
-        required: true
-    get:
-      summary: Your GET endpoint
-      tags: []
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/Post'
-      operationId: userPosts
-components:
-  schemas:
-    Post:
-      title: Post
-      type: object
-      properties:
-        id:
-          type: number
-        userId:
-          type: number
-        title:
-          type: string
-        body:
-          type: string
-    Comment:
-      title: Comment
-      type: object
-      properties:
-        id:
-          type: number
-        name:
-          type: string
-        email:
-          type: string
-        body:
-          type: string
-        postId:
-          type: number
-    User:
-      title: User
-      type: object
-      properties:
-        id:
-          type: number
-        name:
-          type: string
-        username:
-          type: string
-        email:
-          type: string
-        address:
-          $ref: '#/components/schemas/Address'
-        phone:
-          type: string
-        website:
-          type: string
-        company:
-          $ref: '#/components/schemas/Company'
-    Company:
-      title: Company
-      type: object
-      properties:
-        name:
-          type: string
-        catchPhrase:
-          type: string
-        bs:
-          type: string
-    Geo:
-      title: Geo
-      type: object
-      properties:
-        lat:
-          type: string
-        lng:
-          type: string
-    Address:
-      title: Address
-      type: object
-      properties:
-        street:
-          type: string
-        suite:
-          type: string
-        city:
-          type: string
-        zipcode:
-          type: string
-        geo:
-          $ref: '#/components/schemas/Geo'
