--- packages/vite-testapp/.wundergraph/operations/FakeWeather.graphql
@@ -1,12 +1,0 @@
-{
-	getCityByName: weather_getCityByName(name: "Berlin") {
-		id
-		name
-		weather {
-			summary {
-				title
-				description
-			}
-		}
-	}
-}
