--- packages/vite-testapp/.wundergraph/operations/Hello.graphql
@@ -1,3 +1,0 @@
-query ($hello: String!) {
-	gql_hello(name: $hello)
-}
