--- packages/vite-testapp/.wundergraph/operations/ProtectedWeather.graphql
@@ -1,25 +1,0 @@
-query ($forCity: String!) @rbac(requireMatchAll: [admin]) {
-	getCityByName: weather_getCityByName(name: $forCity) {
-		name
-		coord {
-			lat
-			lon
-		}
-		weather {
-			summary {
-				title
-				description
-				icon
-			}
-			temperature {
-				actual
-				min
-				max
-			}
-			wind {
-				speed
-				deg
-			}
-		}
-	}
-}
