--- packages/vite-testapp/.wundergraph/operations/SetName.graphql
@@ -1,3 +1,0 @@
-mutation ($name: String!) {
-	gql_setName(name: $name)
-}
