--- packages/vite-testapp/.wundergraph/operations/Test.graphql
@@ -1,6 +1,0 @@
-query ($input: [gql_TestEnum]!) {
-	gql_testField(inputEnum: $input) {
-		enumList
-		stringList
-	}
-}
