--- packages/vite-testapp/.wundergraph/operations/Weather.graphql
@@ -1,25 +1,0 @@
-query ($forCity: String!) {
-	getCityByName: weather_getCityByName(name: $forCity) {
-		name
-		coord {
-			lat
-			lon
-		}
-		weather {
-			summary {
-				title
-				description
-				icon
-			}
-			temperature {
-				actual
-				min
-				max
-			}
-			wind {
-				speed
-				deg
-			}
-		}
-	}
-}
