--- packages/vite-testapp/.wundergraph/wundergraph.operations.ts
@@ -1,46 +1,0 @@
-import { configureWunderGraphOperations } from '@wundergraph/sdk';
-import type { OperationsConfiguration } from './generated/wundergraph.operations';
-
-export default configureWunderGraphOperations<OperationsConfiguration>({
-	operations: {
-		defaultConfig: {
-			authentication: {
-				required: false,
-			},
-		},
-		queries: (config) => ({
-			...config,
-			caching: {
-				enable: false,
-				staleWhileRevalidate: 60,
-				maxAge: 60,
-				public: true,
-			},
-			liveQuery: {
-				enable: true,
-				pollingIntervalSeconds: 1,
-			},
-		}),
-		mutations: (config) => ({
-			...config,
-		}),
-		subscriptions: (config) => ({
-			...config,
-		}),
-		custom: {
-			ProtectedWeather: (config) => ({
-				...config,
-				authentication: {
-					required: true,
-				},
-			}),
-			PastLaunches: (config) => ({
-				...config,
-				caching: {
-					...config.caching,
-					enable: true,
-				},
-			}),
-		},
-	},
-});
