--- packages/vite-testapp/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "vite-testapp",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "clean": "rimraf ./.next",
-    "dev": "concurrently \"npm run dev:vite\" \"npm run wundergraph\" \"npm run browser\"",
-    "wundergraph": "wunderctl up --debug",
-    "generate": "wunderctl generate",
-    "browser": "wait-on \"http-get://localhost:5173\" && wait-on \"http-get://localhost:9991\" && open-cli http://localhost:5173",
-    "dev:vite": "vite --force",
-    "check": "tsc --noEmit",
-    "build": "pnpm generate && pnpm check && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@wundergraph/sdk": "workspace:*",
-    "@wundergraph/swr": "workspace:*",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "@vitejs/plugin-react": "^2.0.0",
-    "typescript": "^4.6.4",
-    "vite": "^3.0.0",
-    "concurrently": "^6.0.0",
-    "graphql": "^16.3.0",
-    "open-cli": "^7.0.1",
-    "wait-on": "^6.0.0"
-  }
-}
